Mask_RCNN/requirements_bak.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu==1.8
# tensorflow==1.8
keras==2.1.6
opencv-python
h5py
imgaug
IPython[all]

#tensorflow==1.6.0
#keras== 2.2.0

# 必须具有Visual C ++ 2015生成工具
# pip install git+https://github.com/philferriere/cocoapi.git#subdirectory=PythonAPI
# python setup.py build_ext install
# For Python, run "make" under coco/PythonAPI

